trying to fix Unix stat

+++ code/Infrastructure/Verticals/Files/IHardlinkFileService.cs
@@ -2,5 +2,5 @@
 
 public interface IHardlinkFileService
 {
-    uint GetHardLinkCount(string filePath);
+    ulong GetHardLinkCount(string filePath);
 }
